+++ tests/unit/payments/appStoreSubscriptionsService.test.ts
@@ -37,23 +37,127 @@ describe('App Store Subscriptions Service', () => {
 
   it('should get a list of App Store subscriptions', async () => {
     nock(`${host}${PAYMENTS_BASE}`)
-      .get('/appStore/subscriptions')
+      .get('/appStore/subscriptions/')
       .reply(200, createPagedResponse(appStoreSubscription));
 
-    const res = await sdk.payments.appStoreSubscriptions.getSubscriptions();
+    const res = await sdk.payments.appStoreSubscriptions.subscriptions.find();
 
     expect(res.data.length).toBeGreaterThan(0);
   });
 
+  it('should request a list of all App Store subscriptions', async () => {
+    nock(`${host}${PAYMENTS_BASE}`)
+      .get('/appStore/subscriptions/?limit(50)')
+      .reply(200, {
+        page: {
+          total: 65,
+          offset: 0,
+          limit: 50,
+        },
+        data: Array(50).fill(appStoreSubscription),
+      })
+      .get('/appStore/subscriptions/?limit(50,50)')
+      .reply(200, {
+        page: {
+          total: 65,
+          offset: 50,
+          limit: 50,
+        },
+        data: Array(15).fill(appStoreSubscription),
+      });
+    const res =
+      await sdk.payments.appStoreSubscriptions.subscriptions.findAll();
+    expect(res.length).toBe(65);
+  });
+
+  it('should request a list of all App Store subscriptions via iterator', async () => {
+    nock(`${host}${PAYMENTS_BASE}`)
+      .get('/appStore/subscriptions/?limit(50)')
+      .reply(200, {
+        page: {
+          total: 55,
+          offset: 0,
+          limit: 50,
+        },
+        data: Array(50).fill(appStoreSubscription),
+      })
+      .get('/appStore/subscriptions/?limit(50,50)')
+      .reply(200, {
+        page: {
+          total: 55,
+          offset: 50,
+          limit: 50,
+        },
+        data: Array(5).fill(appStoreSubscription),
+      });
+    const subscriptions =
+      sdk.payments.appStoreSubscriptions.subscriptions.findAllIterator();
+
+    await subscriptions.next();
+    const thirdPage = await subscriptions.next();
+    expect(thirdPage.value.data.length).toBe(5);
+  });
+
   it('should get a list of configured App Store subscription products', async () => {
     nock(`${host}${PAYMENTS_BASE}`)
-      .get('/appStore/subscriptions/products')
+      .get('/appStore/subscriptions/products/')
       .reply(200, createPagedResponse(appStoreSubscriptionProduct));
 
-    const res =
-      await sdk.payments.appStoreSubscriptions.getSubscriptionsProducts();
+    const res = await sdk.payments.appStoreSubscriptions.products.find();
 
     expect(res.data.length).toBeGreaterThan(0);
+  });
+
+  it('should request a list of all App Store subscription products', async () => {
+    nock(`${host}${PAYMENTS_BASE}`)
+      .get('/appStore/subscriptions/products/?limit(50)')
+      .reply(200, {
+        page: {
+          total: 65,
+          offset: 0,
+          limit: 50,
+        },
+        data: Array(50).fill(appStoreSubscriptionProduct),
+      })
+      .get('/appStore/subscriptions/products/?limit(50,50)')
+      .reply(200, {
+        page: {
+          total: 65,
+          offset: 50,
+          limit: 50,
+        },
+        data: Array(15).fill(appStoreSubscriptionProduct),
+      });
+    const res = await sdk.payments.appStoreSubscriptions.products.findAll();
+    expect(res.length).toBe(65);
+  });
+
+  it('should request a list of all App Store subscription products via iterator', async () => {
+    nock(`${host}${PAYMENTS_BASE}`)
+      .get('/appStore/subscriptions/products/?limit(50)')
+      .reply(200, {
+        page: {
+          total: 55,
+          offset: 0,
+          limit: 50,
+        },
+        data: Array(50).fill(appStoreSubscriptionProduct),
+      })
+      .get('/appStore/subscriptions/products/?limit(50,50)')
+      .reply(200, {
+        page: {
+          total: 55,
+          offset: 50,
+          limit: 50,
+        },
+        data: Array(5).fill(appStoreSubscriptionProduct),
+      });
+    const products =
+      sdk.payments.appStoreSubscriptions.products.findAllIterator();
+
+    await products.next();
+    const thirdPage = await products.next();
+    expect(thirdPage.value.data.length).toBe(5);
   });
 
   it('should create an App Store subscription product', async () => {
@@ -62,7 +166,7 @@ describe('App Store Subscriptions Service', () => {
       .reply(200, appStoreSubscriptionProduct);
 
     const subscriptionProduct =
-      await sdk.payments.appStoreSubscriptions.createSubscriptionsProduct({
+      await sdk.payments.appStoreSubscriptions.products.create({
         name: 'FibriCheck Premium Monthly',
         appStoreAppBundleId: 'com.qompium.fibricheck',
         appStoreProductId: 'fibricheck-premium-monthly',
@@ -82,10 +186,9 @@ describe('App Store Subscriptions Service', () => {
         affectedRecords: 1,
       });
 
-    const res =
-      await sdk.payments.appStoreSubscriptions.removeSubscriptionsProduct(
-        appStoreSubscriptionProductId
-      );
+    const res = await sdk.payments.appStoreSubscriptions.products.remove(
+      appStoreSubscriptionProductId
+    );
 
     expect(res.affectedRecords).toBe(1);
   });
@@ -97,13 +200,12 @@ describe('App Store Subscriptions Service', () => {
         affectedRecords: 1,
       });
 
-    const res =
-      await sdk.payments.appStoreSubscriptions.updateSubscriptionsProduct(
-        appStoreSubscriptionProductId,
-        {
-          name: 'FibriCheck Premium Monthly',
-        }
-      );
+    const res = await sdk.payments.appStoreSubscriptions.products.update(
+      appStoreSubscriptionProductId,
+      {
+        name: 'FibriCheck Premium Monthly',
+      }
+    );
 
     expect(res.affectedRecords).toBe(1);
   });